Enemies base done

+++ Assets/_Scripts/Enemies/AI_FSM/Tasks/EnemyAttack.cs
@@ -3,7 +3,7 @@ using AI_FSM;
 
 public abstract class EnemyAttack : TaskBase
 {
-    [SerializeField, Range(0f,100f)] protected float _damage;
+    [SerializeField, Range(0f,100f)] protected int _damage;
     [SerializeField, Range(0f, 100f)] protected float _range;
     [SerializeField, Range(0f, 100f)] protected float _attackArea;
